+++ buildSrc/src/main/kotlin/scripts/infrastructure.gradle.kts
@@ -19,12 +19,11 @@
 package scripts
 
 import findVersion
-import org.gradle.configurationcache.extensions.capitalized
-import scripts.Variants_gradle.Default
+import uppercaseFirstChar
 import java.util.Properties
 
 tasks.register("clean", Delete::class) {
-    delete(rootProject.buildDir)
+    delete(rootProject.layout.buildDirectory)
 }
 
 tasks.named<Wrapper>("wrapper") {
@@ -37,8 +36,8 @@ tasks.register("runUnitTests") {
     dependsOn(":app:test${Default.BUILD_VARIANT}UnitTest")
     // kalium have only 2 build type debug or release
     val buildType =
-        if (Default.resolvedBuildType() == Variants_gradle.BuildTypes.DEBUG) Default.resolvedBuildType().capitalized()
-        else Variants_gradle.BuildTypes.RELEASE.capitalized()
+        if (Default.resolvedBuildType() == BuildTypes.DEBUG) Default.resolvedBuildType().uppercaseFirstChar()
+        else BuildTypes.RELEASE.uppercaseFirstChar()
 
     // valid submodules path to run unit tests
     val validSubprojects = setOf("core", "features")
